% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COADrnagn}
\alias{COADrnagn}
\title{COADrnagn: TCGA RNASeq2GeneNorm expression data in HSDS}
\format{SummarizedExperiment with DelayedArray interface to HSDS}
\usage{
COADrnagn
}
\description{
COADrnagn: TCGA RNASeq2GeneNorm expression data in HSDS
}
\note{
Data retrieved using curatedTCGAData 13 Jan 2019 from lwaldronlab
}
\examples{
COADrnagn
}
\keyword{datasets}
