% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprByMSI.R
\name{exprByMSI}
\alias{exprByMSI}
\title{bind microsatellite instability scores via MSIsensor to a TCGA expression dataset}
\usage{
exprByMSI(tumcode, genesym, alias)
}
\arguments{
\item{tumcode}{character(1) e.g., "COAD"}

\item{genesym}{character(1) gene symbol}

\item{alias}{character(1) an alternative name for gene}
}
\description{
bind microsatellite instability scores via MSIsensor to a TCGA expression dataset
}
\note{
This function works specifically with the *rnagn (RESTful)
SummarizedExperiment instance in package tumex5.  The assay
data are in hsdshdflab.hdfgroup.org and internet access
is require for this function to succeed.
}
